Do not claim stall recovery from a reset that leaves the barrier gating
terminalOutputDidReset resolved every open stall episode before
checking whether an existing replay barrier survives the reset. In the
barrier-active non-ack branch the reset only marks owed output and
returns with the barrier still dropping frames, so a still-frozen
terminal was reported as recovered and the eventual real barrier clear
lost its duration. The resolution now runs only on the paths that
actually replace or clear the surface's gate state, with a regression
test covering the barrier-still-active branch.

The reset and self-heal replay entrypoints move to a new
MobileShellComposite+TerminalOutputReset.swift so the delivery file
stays inside the length budget.

Part of #7573
